refactor(chat): split Tab god-file into settings, DOM, and scroll modules
Extract provider/draft settings, tab shell DOM construction, and auto-scroll
helpers from Tab.ts into focused modules. Tab.ts keeps lifecycle and
initialization, re-exporting the public settings API for existing callers.
